requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
deprecation==2.1.0
et-xmlfile==1.1.0
Flask==2.2.2
fonttools<=4.41.1
fpdf2==2.7.4
img2pdf==0.4.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy<=1.25.2
openpyxl==3.1.2
packaging==23.1
pandas<=2.0.3
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug<=2.3.7
